refactor(types): migrate errors to types module
