--- errors/go.mod
@@ -1,15 +1,0 @@
-module github.com/regen-network/regen-ledger/errors
-
-go 1.18
-
-require (
-	cosmossdk.io/errors v1.0.0-beta.7
-	google.golang.org/grpc v1.51.0
-)
-
-require (
-	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013 // indirect
-	google.golang.org/protobuf v1.27.1 // indirect
-)
