+++ x/data/go.mod
@@ -15,7 +15,6 @@ require (
 	github.com/grpc-ecosystem/grpc-gateway v1.16.0
 	github.com/regen-network/gocuke v0.6.2
 	github.com/regen-network/regen-ledger/api v0.8.0
-	github.com/regen-network/regen-ledger/errors v0.0.0-00010101000000-000000000000
 	github.com/regen-network/regen-ledger/types v1.0.0
 	github.com/spf13/cobra v1.6.1
 	github.com/stretchr/testify v1.8.1
@@ -174,5 +173,3 @@ replace github.com/gogo/protobuf => github.com/regen-network/protobuf v1.3.3-alp
 replace github.com/regen-network/regen-ledger/types => ../../types
 
 replace github.com/regen-network/regen-ledger/api => ../../api
-
-replace github.com/regen-network/regen-ledger/errors => ../../errors
