+++ x/data/server/query_anchor_by_hash.go
@@ -3,8 +3,8 @@ package server
 import (
 	"context"
 
-	regenerrors "github.com/regen-network/regen-ledger/errors"
 	"github.com/regen-network/regen-ledger/types"
+	regenerrors "github.com/regen-network/regen-ledger/types/errors"
 
 	"github.com/regen-network/regen-ledger/x/data"
 )
