+++ x/data/server/query_convert_hash_to_iri.go
@@ -3,7 +3,7 @@ package server
 import (
 	"context"
 
-	regenerrors "github.com/regen-network/regen-ledger/errors"
+	regenerrors "github.com/regen-network/regen-ledger/types/errors"
 	"github.com/regen-network/regen-ledger/x/data"
 )
