+++ x/data/server/query_convert_iri_to_hash.go
@@ -3,7 +3,7 @@ package server
 import (
 	"context"
 
-	regenerrors "github.com/regen-network/regen-ledger/errors"
+	regenerrors "github.com/regen-network/regen-ledger/types/errors"
 	"github.com/regen-network/regen-ledger/x/data"
 )
