+++ x/ecocredit/base/keeper/query_class_creator_allowlist.go
@@ -3,7 +3,7 @@ package keeper
 import (
 	"context"
 
-	regenerrors "github.com/regen-network/regen-ledger/errors"
+	regenerrors "github.com/regen-network/regen-ledger/types/errors"
 	types "github.com/regen-network/regen-ledger/x/ecocredit/base/types/v1"
 )
