+++ x/ecocredit/basket/keeper/query_balance.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/cosmos/cosmos-sdk/orm/types/ormerrors"
-	regenerrors "github.com/regen-network/regen-ledger/errors"
+	regenerrors "github.com/regen-network/regen-ledger/types/errors"
 
 	types "github.com/regen-network/regen-ledger/x/ecocredit/basket/types/v1"
 )
